+++ src/main/resources/static/html/size.html
@@ -1,0 +1,69 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet"
+    integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
+<link rel="stylesheet" href="./style.css">
+    <title>Document</title>
+</head>
+<body>
+    <nav class="navbar navbar-expand-lg navbar-light bg-light">
+        <a class="navbar-brand" href="#">Navbar</a>
+        <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
+          <span class="navbar-toggler-icon"></span>
+        </button>
+        <div class="collapse navbar-collapse" id="navbarNav">
+            <ul class="navbar-nav">
+              <li class="nav-item active">
+                <a class="nav-link" href="../index.html">Home <span class="sr-only">(current)</span></a>
+              </li>
+                  <li class="nav-item active">
+                    <a class="nav-link" href="./collection.html">View Collection</a>
+                  </li>
+                  <li class="nav-item active">
+                    <a class="nav-link" href="./artist.html">Find Artist</a>
+                  </li>
+                  <li class="nav-item active">
+                    <a class="nav-link" href="./album.html">Find Album</a>
+                  </li>
+                  <li class="nav-item active">
+                    <a class="nav-link" href="./size.html">Find Vinyl by Size</a>
+                  </li>
+                  <li class="nav-item active">
+                    <a class="nav-link" href="./id.html">Find Vinyl by ID</a>
+                  </li>      
+      <li class="nav-item active">
+        <a class="nav-link" href="./addition.html">Add Vinyl</a>
+      </li>
+      <li class="nav-item active">
+        <a class="nav-link" href="./update.html">Update Vinyl</a>
+      </li>
+      <li class="nav-item active">
+        <a class="nav-link" href="./delete.html">Delete Vinyl</a>
+      </li>
+    </ul>
+</div>
+</nav>
+<div class="card">
+  <form>
+      <h2>Find Size</h2>
+      <input type="number" id="inputFindSize" class="form-control mt-2" placeholder="Vinyl Size" />
+      <button type="button" class="btn btn-primary mt-2" id="buttonSize">Find Size</button>
+      <ul id="sizeOutput"></ul>
+  </form>
+</div>
+<div class="card2">
+  <form>
+      <ul id="listOutput"></ul>
+  </form>
+</div>
+<script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script>
+<script type="module" src="../js/size.js"></script>
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js"
+    integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p"
+    crossorigin="anonymous"></script>
+</body>
+</html>
