+++ src/main/resources/static/index.html
@@ -1,86 +1,67 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet"
-    integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
-<link rel="stylesheet" href="./style.css">
+        integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
+    <link rel="stylesheet" href="./style.css">
     <title>Vinyl Collection</title>
 </head>
+
 <body>
-    <div>
-        <div>
-            <form>
-                <h2>Read All</h2>
-                <button type="button" class="btn" id="buttonReadAll">Read All</button>
-                <ul id="listOutput"></ul>
-            </form>
+    <nav class="navbar navbar-expand-lg navbar-light bg-light">
+        <a class="navbar-brand" href="#">Navbar</a>
+        <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
+          <span class="navbar-toggler-icon"></span>
+        </button>
+        <div class="collapse navbar-collapse" id="navbarNav">
+            <ul class="navbar-nav">
+              <li class="nav-item active">
+                <a class="nav-link" href="./index.html">Home <span class="sr-only">(current)</span></a>
+              </li>
+                  <li class="nav-item active">
+                    <a class="nav-link" href="html/collection.html">View Collection</a>
+                  </li>
+                  <li class="nav-item active">
+                    <a class="nav-link" href="html/artist.html">Find Artist</a>
+                  </li>
+                  <li class="nav-item active">
+                    <a class="nav-link" href="html/album.html">Find Album</a>
+                  </li>
+                  <li class="nav-item active">
+                    <a class="nav-link" href="html/size.html">Find Vinyl by Size</a>
+                  </li>
+                  <li class="nav-item active">
+                    <a class="nav-link" href="html/id.html">Find Vinyl by ID</a>
+                  </li>   
+      <li class="nav-item active">
+        <a class="nav-link" href="html/addition.html">Add Vinyl</a>
+      </li>
+      <li class="nav-item active">
+        <a class="nav-link" href="html/update.html">Update Vinyl</a>
+      </li>
+      <li class="nav-item active">
+        <a class="nav-link" href="html/delete.html">Delete Vinyl</a>
+      </li>
+    </ul>
+</div>
+</nav>
+<div>
+    <p>
+        Welcome to your vinyl colleciton. Please select an option above to either view or modify your collection!
+    </p>
+    <h2>Happy Spinning</h2>
+</div>
+            <script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script>
+            <script type="module" src="./js/script.js"></script>
+            <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js"
+                integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p"
+                crossorigin="anonymous"></script>
         </div>
-        <div>
-            <form>
-                <h2>Create</h2>
-                <input type="text" id="inputArtist" class="form-control" placeholder="Artist Name"/>
-                <input type="text" id="inputAlbum" class="form-control" placeholder="Album Name"/>
-                <input type="number" id="inputSize" class="form-control" placeholder="Vinyl Size"/>
-                <button type="button" class="btn btn-primary mt-2" id="buttonCreate">Create</button>
-            </form>
-        </div>
-        <div>
-            <form>
-                <h2>Find Vinyl By ID</h2>
-                <input type="number" id="inputId" class="form-control mt-2" placeholder="Item ID" />
-                <button type="button" class="btn btn-primary mt-2" id="buttonRead">ID READ</button>
-                <ul id="newOutput"></ul>
-            </form>
-        </div>
-        <div>
-            <form>
-                <h2>Find artist</h2>
-                <input type="text" id="inputFindArtist" class="form-control mt-2" placeholder="Artist"/>
-                <button type="button" class="btn btn-primary mt-2" id="buttonArtist">Find Artist</button>
-                <ul id="artistOutput"></ul>
-            </form>
-        </div>
-        <div>
-            <form>
-                <h2>Find Album</h2>
-                <input type="text" id="inputFindAlbum" class="form-control mt-2" placeholder="Album"/>
-                <button type="button" class="btn btn-primary mt-2" id="buttonAlbum">Find Album</button>
-                <ul id="albumOutput"></ul>
-            </form>
-        </div>
-        <div>
-            <form>
-                <h2>Find Size</h2>
-                <input type="number" id="inputFindSize" class="form-control mt-2" placeholder="Vinyl Size" />
-                <button type="button" class="btn btn-primary mt-2" id="buttonSize">Find Size</button>
-                <ul id="sizeOutput"></ul>
-            </form>
-        </div>
-        <div>
-            <form>
-                <h2>Update Vinyl</h2>
-                <input type="number" id="inputUpdateId" class="form-control mt-2" placeholder="Vinyl ID"/>
-                <input type="text" id="inputUpdateArtist" class="form-control mt-2" placeholder="Artist Name"/>
-                <input type="text" id="inputUpdateAlbum" class="form-control mt-2" placeholder="Album Name"/>
-                <input type="number" id="inputUpdateSize" class="form-control mt-2" placeholder="Vinyl Size"/>
-                <button type="button" class="btn btn-primary mt-2" id="buttonUpdate">Update</button>
-            </form>
-        </div>
-        <div>
-            <form>
-                <h2>Delete Vinyl</h2>
-                <input type="number" id="inputDeleteId" class="form-control mt-2" placeholder="Vinyl ID" />
-                <button type="button" class="btn btn-primary mt-2" id="buttonDelete">Delete</button>
-            </form>
-        </div>
-        <script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script>
-        <script type="module" src="./script.js"></script>
-        <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js"
-            integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p"
-            crossorigin="anonymous"></script>
     </div>
 </body>
+
 </html>
