+++ src/main/resources/static/style.css
@@ -1,0 +1,3 @@
+body {
+    background-color: grey;
+}
